Package update
Update to use org.openalpr package name in Jave instead of com.openalpr.jni. This is to make Java JNI package compatible to Android version.

Note that to use the library with Sapphire in Android, generated library file (/usr/lib/openalprjni.so) should be renamed to the same library name as in Android openalpr-native.so since Android code uses openalpr-native for loading JNI library.

TODO: generate .so file as openalpr-native.so instead of openalprjni.so