Support processing at device and server in the same process.
1.  Add region information registration to distinguish processing between device and server.
2. Processing time state display in UI.
3. Support Amazon Kindle.

Note it assumes OMS is running on the same network. To run the computation at Cloud, OMS should run on Cloud and the app will connect to OMS in the cloud. To run the processing on the device, OMS should run on the same network as the device (or device should be reachable from OMS).