2018-4-6-ObjMigrationFromDeviceToServerViaTwoOms
Object migration from device to server via two different OMSs is complete.
The scenarios is this:
Device is in an internal network, hence, uses its own OMS where OMS can reach the deivce - OMS1.
Server is in a different network where it cannot reach the device (device to server is reachable). Example of this scenario is cloud server. OMS is launched separately to support Sapphire kernel servers running on the cloud server - OMS2.

This check-in enables pushing object on the device Sapphire kernel to server side Sapphire kernel. Internally, when any method name containing "migrateObjectToDifferentOMS" is called, it will push the object registered in OMS1 to the server registered in OMS2. OMS1 will lose the registered object and OMS2 will have a newly registered object. Therefore, state is preserved between the two entities (device vs server). Here, the state is count of each license plate reading.